+++ Teensy/lib/PinneRobot/PinneMotor.h
@@ -89,14 +89,14 @@ class PinneMotor
           void GoToParkingPosition();
           void GoToTargetPositionByDuration(int targetPosition, int duration,
                                             double minSpeed, double beta,
-                                            double skirtRatio, uint8_t moveId);
+                                            double skirtRatio, int32_t moveId);
           void GoToTargetPositionByMaxSpeed(int targetPosition, double maxSpeed,
                                             double minSpeed, double beta,
-                                            double skirtRatio, uint8_t moveId);
+                                            double skirtRatio, int32_t moveId);
           void GoToTargetPositionByConstantSpeed(int targetPosition,
                                                  double speed, double minSpeed,
                                                  double beta,
-                                                 double skirtRatio, uint8_t moveId);
+                                                 double skirtRatio, int32_t moveId);
 
           bool routeOSC(OSCMessage &msg, int initialOffset);
           uint8_t GetBlockingMask() { return _blockingMask; }
